+++ Brainarr.Plugin/Models/Providers/AnthropicModels.cs
@@ -1,0 +1,53 @@
+using System.Collections.Generic;
+using System.Text.Json.Serialization;
+
+namespace Brainarr.Plugin.Models.Providers
+{
+    /// <summary>
+    /// Anthropic Claude response format models
+    /// </summary>
+    public class AnthropicResponse
+    {
+        [JsonPropertyName("id")]
+        public string Id { get; set; }
+
+        [JsonPropertyName("type")]
+        public string Type { get; set; }
+
+        [JsonPropertyName("role")]
+        public string Role { get; set; }
+
+        [JsonPropertyName("content")]
+        public List<AnthropicContent> Content { get; set; }
+
+        [JsonPropertyName("model")]
+        public string Model { get; set; }
+
+        [JsonPropertyName("stop_reason")]
+        public string StopReason { get; set; }
+
+        [JsonPropertyName("stop_sequence")]
+        public string StopSequence { get; set; }
+
+        [JsonPropertyName("usage")]
+        public AnthropicUsage Usage { get; set; }
+    }
+
+    public class AnthropicContent
+    {
+        [JsonPropertyName("type")]
+        public string Type { get; set; }
+
+        [JsonPropertyName("text")]
+        public string Text { get; set; }
+    }
+
+    public class AnthropicUsage
+    {
+        [JsonPropertyName("input_tokens")]
+        public int InputTokens { get; set; }
+
+        [JsonPropertyName("output_tokens")]
+        public int OutputTokens { get; set; }
+    }
+}
