+++ Brainarr.Plugin/Models/Providers/GeminiModels.cs
@@ -1,0 +1,62 @@
+using System.Collections.Generic;
+using System.Text.Json.Serialization;
+
+namespace Brainarr.Plugin.Models.Providers
+{
+    /// <summary>
+    /// Google Gemini response format models
+    /// </summary>
+    public class GeminiResponse
+    {
+        [JsonPropertyName("candidates")]
+        public List<GeminiCandidate> Candidates { get; set; }
+
+        [JsonPropertyName("promptFeedback")]
+        public GeminiPromptFeedback PromptFeedback { get; set; }
+    }
+
+    public class GeminiCandidate
+    {
+        [JsonPropertyName("content")]
+        public GeminiContent Content { get; set; }
+
+        [JsonPropertyName("finishReason")]
+        public string FinishReason { get; set; }
+
+        [JsonPropertyName("index")]
+        public int Index { get; set; }
+
+        [JsonPropertyName("safetyRatings")]
+        public List<GeminiSafetyRating> SafetyRatings { get; set; }
+    }
+
+    public class GeminiContent
+    {
+        [JsonPropertyName("parts")]
+        public List<GeminiPart> Parts { get; set; }
+
+        [JsonPropertyName("role")]
+        public string Role { get; set; }
+    }
+
+    public class GeminiPart
+    {
+        [JsonPropertyName("text")]
+        public string Text { get; set; }
+    }
+
+    public class GeminiSafetyRating
+    {
+        [JsonPropertyName("category")]
+        public string Category { get; set; }
+
+        [JsonPropertyName("probability")]
+        public string Probability { get; set; }
+    }
+
+    public class GeminiPromptFeedback
+    {
+        [JsonPropertyName("safetyRatings")]
+        public List<GeminiSafetyRating> SafetyRatings { get; set; }
+    }
+}
